p_enemy: don't sign-extend MBF21 flag args in the flag codepointers
A_AddFlags, A_RemoveFlags and A_JumpIfFlagsSet read their flag arguments
from state->args[] (a 32-bit signed long on LLP64 targets such as
MinGW64) and cast them straight to uint64_t:

    flags = (uint64_t)actor->state->args[0];

A flag value with bit 31 set is negative as a 32-bit long, so the cast
sign-extends it to 0xFFFFFFFF80000000 -- smearing ones across the whole
high 32 bits of the 64-bit flag word.  A_AddFlags then ORs that in,
setting a pile of bogus high flags (e.g. the bit-63 internal flags);
A_RemoveFlags ANDs with its complement, clearing a swath of unrelated
flags; A_JumpIfFlagsSet compares against a corrupted mask.

dsda-doom does not hit this because its args[] is long long (64-bit), so
bit 31 stays positive.  This core's args[] is long, so cast through
uint32_t first to take the low 32 bits without sign extension:

    flags = (uint64_t)(uint32_t)actor->state->args[0];

Found by a proactive audit of the MBF21 codepointer set; the rest
(damage-modulo guards, spawn null-checks, bounded state transitions)
were already sound.

+++ src/p_enemy.c
@@ -3158,8 +3158,11 @@ void A_JumpIfFlagsSet(mobj_t *actor)
   if (!mbf21_features || !actor)
     return;
   state  = actor->state->args[0];
-  flags  = (uint64_t)actor->state->args[1];
-  flags2 = (uint64_t)actor->state->args[2];
+  /* args[] is a 32-bit (signed) long on LLP64 targets; cast through
+   * uint32_t first so a flag value with bit 31 set is not sign-extended
+   * into the high 32 bits of the 64-bit flag word. */
+  flags  = (uint64_t)(uint32_t)actor->state->args[1];
+  flags2 = (uint64_t)(uint32_t)actor->state->args[2];
   if ((actor->flags & flags) == flags &&
       (actor->flags2 & flags2) == flags2)
     P_SetMobjState(actor, state);
@@ -3171,8 +3174,9 @@ void A_AddFlags(mobj_t *actor)
   dbool update_blockmap;
   if (!mbf21_features || !actor)
     return;
-  flags  = (uint64_t)actor->state->args[0];
-  flags2 = (uint64_t)actor->state->args[1];
+  /* see A_JumpIfFlagsSet: avoid sign-extending a bit-31 flag value */
+  flags  = (uint64_t)(uint32_t)actor->state->args[0];
+  flags2 = (uint64_t)(uint32_t)actor->state->args[1];
   update_blockmap = ((flags & MF_NOBLOCKMAP) && !(actor->flags & MF_NOBLOCKMAP))
                  || ((flags & MF_NOSECTOR)   && !(actor->flags & MF_NOSECTOR));
   if (update_blockmap)
@@ -3189,8 +3193,9 @@ void A_RemoveFlags(mobj_t *actor)
   dbool update_blockmap;
   if (!mbf21_features || !actor)
     return;
-  flags  = (uint64_t)actor->state->args[0];
-  flags2 = (uint64_t)actor->state->args[1];
+  /* see A_JumpIfFlagsSet: avoid sign-extending a bit-31 flag value */
+  flags  = (uint64_t)(uint32_t)actor->state->args[0];
+  flags2 = (uint64_t)(uint32_t)actor->state->args[1];
   update_blockmap = ((flags & MF_NOBLOCKMAP) && (actor->flags & MF_NOBLOCKMAP))
                  || ((flags & MF_NOSECTOR)   && (actor->flags & MF_NOSECTOR));
   if (update_blockmap)
